Add .ini option SkipUi to skip the dialog
Instead the values used in the .ini file will be used. So ideally you
use the dialog once, make your selections and launch the game so your
configurations is saved. Then you close the game and edit the
legacyaltrenderer.ini to SkipUi = True, and from then on no dialog is
shown anymore, allowing us to skip this obsolete step. To change the
through the UI, set SkipUi = False again so the dialog will show up
again.